requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-bootstrap4==0.0.8
djangorestframework==3.8.2
django-rest-swagger==2.2.0
pytz==2018.5
django-environ==0.4.1